Updated loading and normalization stats

+++ module.py
@@ -26,8 +26,11 @@ class SIGReg(torch.nn.Module):
         """
         proj: (T, B, D)
         """
-        # sample random projections
-        A = torch.randn(proj.size(-1), self.num_proj, device=proj.device)
+        # sample random projections — must match input dtype for bf16-mixed
+        A = torch.randn(
+            proj.size(-1), self.num_proj,
+            device=proj.device, dtype=proj.dtype,
+        )
         A = A.div_(A.norm(p=2, dim=0))
         # compute the epps-pulley statistic
         x_t = (proj @ A).unsqueeze(-1) * self.t
